[Install] Add install rule to Makefile
This install rule is in place in order to copy the binary into /usr/bin
to allow to use the program from any project without the hassle to copy
the binary into the repository.
The user_install rule is quite similar but instead of copying the binary
into /usr/bin, it copies it into ${HOME}/.local/bin for those who want
to install it for a single user instead of installing it system wide.
This requires you to have this path in your PATH environment variable.